First iteration of synchronous validation (creation time)
After trying out different alternatives I decided to go with a
duplicate of the CKAN core resource_create action, with just the
addition of a synchronous data validation step.

This is of course not ideal but it's the only way right now to hook
reliably into the creation process without overcomplicating things.
Hopefully future versions of CKAN will incorporate more flexible hook
points that will allow a better approach.

When `ckanext.validation.run_on_create_sync` is True a validation is
performed and if errors are found a ValidationError is raised with the
report (we will need a custom helper to parse this properly).