Delete a resource that was created during sync validation
A the resource_validation_run and run_validation_job function do a
commit, the resource was stored even if a ValidationError was raised.
This patch explicitly deletes the resource but the whole sync process
needs to be refactored, as this keeps a deleted resource in the
resource table in creates 3 new revisions. We shouldn't touch the
database on sync validation.