Pass Authorization header on private resources
When validating a locally uploaded resource, the extension uses the
actual physical path to read the file, so there's no need for
authorization. But when the upload is on a cloud-based backend (like the
ones provided by ckanext-cloudstorage or ckanext-s3filestore) we need to
request the file via an HTTP request to CKAN. If the resource is private
this will require an Authorization header in order to avoid a Not
Authorized error.

On non-local uploads (ie not implementing ckan.lib.uploader.ResourceUpload)
the API key for the site user will be passed as part of the request (or
alternatively ckanext.validation.pass_auth_header_value if provided).

As this involves sending API keys to other extensions this behaviour can
be turned off by setting ckanext.validation.pass_auth_header to False.