Don't process extra resources on resource_create
In async mode, if a dataset already has one or more resources suitable
for validation, when adding a new one all the previous ones are also
sent to validate.

@ThrawnCA patch in #67, dealt with the same scenario but when updating
an existing resource.

In this case sadly we need to use the context, as there is no way to
reliably tell in `after_update` if it's part of a resource_create call