Add turn_off_validation context manager
Rather than try to deal with complex logic on the plugin hooks with
dodgy context keys, it offers a way to shut down completely the
validation process in a specific block of code. Works by explicitly
setting all the validation config options to False and restoring the
previous values afterwards.